decided to move all the code coming from ai assisted brachy into an isolated folder. in here, we can start deleting things as needed
