+++ brachyutils/geometry/catheter_utils/patch_ai_assisted_brachy/catheter_api.py
@@ -1,10 +1,10 @@
 from pathlib import Path
 from typing import Tuple, Literal, Dict, List
 import SimpleITK as sitk
-from brachyutils.geometry.catheter_utils.catheter_setup import CatheterSetUp
-from brachyutils.geometry.catheter_utils.contour.creator import CatheterContourCreator
-from brachyutils.geometry.catheter_utils.digitization.contour_digitizer import DwellPositionCreator, CatheterTableTimesFiller
-from brachyutils.geometry.catheter_utils.contour.postprocess import CatheterPostProcessor
+from brachyutils.geometry.catheter_utils.patch_ai_assisted_brachy.catheter_setup import CatheterSetUp
+from brachyutils.geometry.catheter_utils.patch_ai_assisted_brachy.contour.creator import CatheterContourCreator
+from brachyutils.geometry.catheter_utils.patch_ai_assisted_brachy.digitization.contour_digitizer import DwellPositionCreator, CatheterTableTimesFiller
+from brachyutils.geometry.catheter_utils.patch_ai_assisted_brachy.contour.postprocess import CatheterPostProcessor
 
 import numpy as np
 import json
