+++ brachyutils/geometry/catheter_utils/patch_ai_assisted_brachy/digitization/spline_interpolator.py
@@ -11,14 +11,14 @@ from scipy.optimize import minimize_scalar
 from scipy.spatial.distance import cdist
 from scipy import ndimage
 
-from brachyutils.geometry.catheter_utils.utils import create_slicer_markup_points, crop_around_mask
-from brachyutils.geometry.catheter_utils.catheter_setup import CatheterSetUp
-from brachyutils.geometry.catheter_utils.digitization.rotation import (
+from brachyutils.geometry.catheter_utils.patch_ai_assisted_brachy.utils import create_slicer_markup_points, crop_around_mask
+from brachyutils.geometry.catheter_utils.patch_ai_assisted_brachy.catheter_setup import CatheterSetUp
+from brachyutils.geometry.catheter_utils.patch_ai_assisted_brachy.digitization.rotation import (
     calculate_rotation_matrix,
     create_rotation_transform,
     rotate_volume
 )
-from brachyutils.geometry.catheter_utils.utils import (
+from brachyutils.geometry.catheter_utils.patch_ai_assisted_brachy.utils import (
     distance,
     find_extremal_points_a,
     fit_line,
