+++ brachyutils/geometry/catheter_utils/patch_ai_assisted_brachy/utils.py
